+++ src/commands/dxcop/source/check.ts
@@ -12,8 +12,10 @@ import { EmailToCaseSettingsRuleset } from '../../../ruleset/emailToCaseSettings
 import { LwcMetadataRuleset } from '../../../ruleset/lwcMetadataRuleset';
 import { MetadataProblem } from '../../../ruleset/metadataProblem';
 import { MetadataRuleset } from '../../../ruleset/metadataRuleset';
+import { MinimumAccessProfileRuleset } from '../../../ruleset/minimumAccessProfileRuleset';
 import { RecordTypePicklistRuleset } from '../../../ruleset/recordTypePicklistRuleset';
 import { RecordTypePicklistValueRuleset } from '../../../ruleset/recordTypePicklistValueRuleset';
+import { QueueRuleset } from '../../../ruleset/queueRuleset';
 
 // Initialize Messages with the current plugin directory
 Messages.importMessagesDirectory(__dirname);
@@ -74,6 +76,14 @@ export default class Check extends SfdxCommand {
     }
     if (config.ruleSets.lightningWebComponents.enabled) {
       rulesets.push(new LwcMetadataRuleset(sfdxProjectBrowser));
+    }
+    if (config.ruleSets.minimumAccessProfile.enabled) {
+      rulesets.push(
+        new MinimumAccessProfileRuleset(sfdxProjectBrowser, config.ruleSets.minimumAccessProfile.profileName)
+      );
+    }
+    if (config.ruleSets.queues.enabled) {
+      rulesets.push(new QueueRuleset(sfdxProjectBrowser));
     }
     if (config.ruleSets.recordTypePicklists.enabled) {
       rulesets.push(new RecordTypePicklistRuleset(sfdxProjectBrowser));